rename: 디렉토리 이름 ErrorBoundary 로 변경
